move code to postgres crate
